+++ designer/server/src/models/forms/editor-v2/advanced-settings-fields.js
@@ -1,5 +1,76 @@
+import {
+  ComponentType,
+  classesSchema,
+  hintTextSchema,
+  maxFutureSchema,
+  maxLengthSchema,
+  maxPastSchema,
+  maxSchema,
+  minLengthSchema,
+  minSchema,
+  nameSchema,
+  precisionSchema,
+  prefixSchema,
+  questionOptionalSchema,
+  questionSchema,
+  questionTypeFullSchema,
+  regexSchema,
+  rowsSchema,
+  shortDescriptionSchema,
+  suffixSchema
+} from '@defra/forms-model'
+import Joi from 'joi'
+
 import { QuestionAdvancedSettings } from '~/src/common/constants/editor.js'
 import { GOVUK_LABEL__M } from '~/src/models/forms/editor-v2/common.js'
+
+export const advancedSettingsPerComponentType =
+  /** @type {Record<string, string[]> } */ ({
+    TextField: [
+      QuestionAdvancedSettings.MinLength,
+      QuestionAdvancedSettings.MaxLength,
+      QuestionAdvancedSettings.Regex,
+      QuestionAdvancedSettings.Classes
+    ],
+    MultilineTextField: [
+      QuestionAdvancedSettings.MinLength,
+      QuestionAdvancedSettings.MaxLength,
+      QuestionAdvancedSettings.Rows,
+      QuestionAdvancedSettings.Regex,
+      QuestionAdvancedSettings.Classes
+    ],
+    YesNoField: [],
+    DatePartsField: [
+      QuestionAdvancedSettings.MaxPast,
+      QuestionAdvancedSettings.MaxFuture,
+      QuestionAdvancedSettings.Classes
+    ],
+    MonthYearField: [
+      QuestionAdvancedSettings.MaxPast,
+      QuestionAdvancedSettings.MaxFuture,
+      QuestionAdvancedSettings.Classes
+    ],
+    SelectField: [],
+    AutocompleteField: [],
+    RadiosField: [],
+    CheckboxesField: [],
+    NumberField: [
+      QuestionAdvancedSettings.Min,
+      QuestionAdvancedSettings.Max,
+      QuestionAdvancedSettings.Precision,
+      QuestionAdvancedSettings.Prefix,
+      QuestionAdvancedSettings.Suffix
+    ],
+    UkAddressField: [],
+    TelephoneNumberField: [QuestionAdvancedSettings.Classes],
+    EmailAddressField: [QuestionAdvancedSettings.Classes],
+    Html: [],
+    InsetText: [],
+    Details: [],
+    List: [],
+    Markdown: [],
+    FileUploadField: []
+  })
 
 /**
  * @type { Record<string, GovukField> }
@@ -146,6 +217,78 @@ export const allAdvancedSettingsFields =
     }
   })
 
+export const baseSchema = Joi.object().keys({
+  name: nameSchema,
+  question: questionSchema.messages({
+    '*': 'Enter a question'
+  }),
+  hintText: hintTextSchema,
+  questionOptional: questionOptionalSchema,
+  shortDescription: shortDescriptionSchema.messages({
+    '*': 'Enter a short description'
+  }),
+  questionType: questionTypeFullSchema.messages({
+    '*': 'The question type is missing'
+  })
+})
+
+export const allSpecificSchemas = Joi.object().keys({
+  maxFuture: maxFutureSchema.messages({
+    '*': 'Max days in the future must be a number greater than zero'
+  }),
+  maxPast: maxPastSchema.messages({
+    '*': 'Max days in the past must be a number greater than zero'
+  }),
+  min: minSchema.messages({
+    '*': 'Lowest number must be a number'
+  }),
+  max: maxSchema.messages({
+    '*': 'Highest number must be a number greater than zero'
+  }),
+  minLength: minLengthSchema.messages({
+    '*': 'Minimum length must be a number greater than zero'
+  }),
+  maxLength: maxLengthSchema.messages({
+    '*': 'Maximum length must be a number greater than zero'
+  }),
+  precision: precisionSchema.messages({
+    '*': 'Precision must be a number greater than zero'
+  }),
+  prefix: prefixSchema,
+  suffix: suffixSchema,
+  regex: regexSchema,
+  rows: rowsSchema.messages({
+    '*': 'Rows must be a number greater than zero'
+  }),
+  classes: classesSchema
+})
+
 /**
- * @import { GovukField } from '@defra/forms-model'
+ * @param {GovukField} field
+ */
+export function getFieldComponentType(field) {
+  switch (field.name) {
+    case QuestionAdvancedSettings.Min:
+    case QuestionAdvancedSettings.Max:
+    case QuestionAdvancedSettings.MinLength:
+    case QuestionAdvancedSettings.MaxLength:
+    case QuestionAdvancedSettings.MaxFuture:
+    case QuestionAdvancedSettings.MaxPast:
+    case QuestionAdvancedSettings.Precision:
+    case QuestionAdvancedSettings.Prefix:
+    case QuestionAdvancedSettings.Rows:
+    case QuestionAdvancedSettings.Suffix:
+      return ComponentType.TextField
+    case QuestionAdvancedSettings.Regex:
+    case QuestionAdvancedSettings.Classes:
+      return ComponentType.MultilineTextField
+    default:
+      throw new Error(
+        `Invalid or not implemented advanced setting field name (${field.name})`
+      )
+  }
+}
+
+/**
+ * @import { ComponentDef, GovukField } from '@defra/forms-model'
  */
