+++ designer/server/src/routes/forms/editor-v2/question-details.js
@@ -1,25 +1,4 @@
-import {
-  classesSchema,
-  hintTextSchema,
-  maxFutureSchema,
-  maxLengthSchema,
-  maxPastSchema,
-  maxSchema,
-  minLengthSchema,
-  minSchema,
-  nameSchema,
-  precisionSchema,
-  prefixSchema,
-  questionOptionalSchema,
-  questionSchema,
-  questionTypeFullSchema,
-  regexSchema,
-  rowsSchema,
-  shortDescriptionSchema,
-  suffixSchema
-} from '@defra/forms-model'
 import { StatusCodes } from 'http-status-codes'
-import Joi from 'joi'
 
 import * as scopes from '~/src/common/constants/scopes.js'
 import { sessionNames } from '~/src/common/constants/session-names.js'
@@ -32,6 +11,10 @@ import { getValidationErrorsFromSession } from '~/src/lib/error-helper.js'
 import * as forms from '~/src/lib/forms.js'
 import { redirectWithErrors } from '~/src/lib/redirect-helper.js'
 import { isCheckboxSelected } from '~/src/lib/utils.js'
+import {
+  allSpecificSchemas,
+  baseSchema
+} from '~/src/models/forms/editor-v2/advanced-settings-fields.js'
 import { CHANGES_SAVED_SUCCESSFULLY } from '~/src/models/forms/editor-v2/common.js'
 import * as viewModel from '~/src/models/forms/editor-v2/question-details.js'
 import { editorv2Path } from '~/src/models/links.js'
@@ -41,54 +24,7 @@ export const ROUTE_FULL_PATH_QUESTION_DETAILS = `/library/{slug}/editor-v2/page/
 
 const errorKey = sessionNames.validationFailure.editorQuestionDetails
 
-export const baseSchema = Joi.object().keys({
-  name: nameSchema,
-  question: questionSchema.messages({
-    '*': 'Enter a question'
-  }),
-  hintText: hintTextSchema,
-  questionOptional: questionOptionalSchema,
-  shortDescription: shortDescriptionSchema.messages({
-    '*': 'Enter a short description'
-  }),
-  questionType: questionTypeFullSchema.messages({
-    '*': 'The question type is missing'
-  })
-})
-
-// To be extended with question specific fields
-const specificsSchema = Joi.object().keys({
-  maxFuture: maxFutureSchema.messages({
-    '*': 'Max days in the future must be a number greater than zero'
-  }),
-  maxPast: maxPastSchema.messages({
-    '*': 'Max days in the past must be a number greater than zero'
-  }),
-  min: minSchema.messages({
-    '*': 'Lowest number must be a number'
-  }),
-  max: maxSchema.messages({
-    '*': 'Highest number must be a number greater than zero'
-  }),
-  minLength: minLengthSchema.messages({
-    '*': 'Minimum length must be a number greater than zero'
-  }),
-  maxLength: maxLengthSchema.messages({
-    '*': 'Maximum length must be a number greater than zero'
-  }),
-  precision: precisionSchema.messages({
-    '*': 'Precision must be a number greater than zero'
-  }),
-  prefix: prefixSchema,
-  suffix: suffixSchema,
-  regex: regexSchema,
-  rows: rowsSchema.messages({
-    '*': 'Rows must be a number greater than zero'
-  }),
-  classes: classesSchema
-})
-
-const schema = baseSchema.concat(specificsSchema)
+const schema = baseSchema.concat(allSpecificSchemas)
 
 /**
  *
@@ -104,12 +40,15 @@ function mapQuestionDetails(payload) {
     options: {
       classes: payload.classes,
       required: !isCheckboxSelected(payload.questionOptional),
-      rows: payload.rows ?? undefined
+      rows: payload.rows ?? undefined,
+      prefix: payload.prefix ?? undefined,
+      suffix: payload.suffix ?? undefined
     },
     schema: {
-      min: payload.minLength ?? undefined,
-      max: payload.maxLength ?? undefined,
-      regex: payload.regex ?? undefined
+      min: payload.minLength ?? payload.min ?? undefined,
+      max: payload.maxLength ?? payload.max ?? undefined,
+      regex: payload.regex ?? undefined,
+      precision: payload.precision ?? undefined
     }
   })
 }
@@ -229,6 +168,6 @@ export default [
 ]
 
 /**
- * @import { FormEditorInputQuestion, ComponentDef } from '@defra/forms-model'
+ * @import { FormEditorInputQuestion, ComponentDef, ComponentType } from '@defra/forms-model'
  * @import { ServerRoute } from '@hapi/hapi'
  */
