+++ model/src/form/form-editor/types.ts
@@ -96,6 +96,31 @@ export interface FormEditor {
    * The classes to be applied to a field
    */
   classes: string
+
+  /**
+   * The prefix to be applied to a field
+   */
+  prefix: string
+
+  /**
+   * The suffix to be applied to a field
+   */
+  suffix: string
+
+  /**
+   * The decimal precision of a number field
+   */
+  precision: string
+
+  /**
+   * The lowest number allowed in a field
+   */
+  min: string
+
+  /**
+   * The highest number allowed in a field
+   */
+  max: string
 }
 
 export type FormEditorInputPage = Pick<
@@ -121,6 +146,11 @@ export type FormEditorInputQuestion = Pick<
   | 'regex'
   | 'rows'
   | 'classes'
+  | 'prefix'
+  | 'suffix'
+  | 'precision'
+  | 'min'
+  | 'max'
 >
 
 export type FormEditorInputPageSettings = Pick<
@@ -151,9 +181,5 @@ export interface FormEditorGovukField {
   hintText?: GovukField
   questionOptional?: GovukField
   shortDescription?: GovukField
-  minLength?: GovukField
-  maxLength?: GovukField
-  regex?: GovukField
-  classes?: GovukField
   errorMessage?: { text: string }
 }
